Feat[#6]: 리뷰 시스템 구현 및 s3 파일 업로드 구현

+++ viewlog/src/main/java/com/trip/viewlog/attraction/domain/Attraction.java
@@ -5,7 +5,7 @@ import lombok.Getter;
 
 @Getter
 public class Attraction {
-	private Integer id;
+	private Long id;
     private Integer contentId;
     private ContentType contentType;
     private Integer areaCode;
@@ -23,7 +23,7 @@ public class Attraction {
     private String overview;
     
     @Builder
-	public Attraction(Integer id, Integer contentId, ContentType contentType, Integer areaCode,
+	public Attraction(Long id, Integer contentId, ContentType contentType, Integer areaCode,
 			Integer siGunGuCode, String title, String firstImage1, String firstImage2, Integer mapLevel,
 			Double latitude, Double longitude, String tel, String addr1, String addr2, String homepage,
 			String overview) {
